+++ Intern/RayCore/src/VulkanEngine/BufferIO/WriteBufferRaw.cpp
@@ -26,7 +26,7 @@ void VulkanEngine::writeBufferRaw(const char* bufname, char* indata) {
     }
 }
 
-void VulkanEngine::storeToStagingBuffer(char* indata, uint32_t bytes) {
+void VulkanEngine::storeToStagingBuffer(char* indata, size_t bytes) {
     void* buf;
     vkMapMemory(m_Device, m_stagingMemory, 0, STAGING_SIZE, 0, &buf);
